根据启发调整了数据同步的菜单、以及增加部分权限
