+++ README.md
@@ -2,16 +2,16 @@
 
 **前言**：
 >
-> 本应用主要基于OdooERP开发，支持社区版和企业版，当前版本仅支持12版本 可自行下载后根据实际情况进行增加完善和调整功能。 **但是我希望能尊重作者的劳动成果，不要改动作者信息。谢谢~** 
+> 本应用主要基于OdooERP开发，支持社区版和企业版，当前版本仅支持12版本 开发者可自行下载后根据实际情况进行增加完善和调整功能。 **但是我希望能尊重作者的劳动成果，不要改动作者信息。谢谢~** 
 > 
-> 在使用本模块前，请先将钉钉中你创建的E应用或微应用权限放开和配置出口ip，得到钉钉应用的 **AppKey**和 **AppSecret**， 至于钉钉后台中的配置请参照：>https://open-doc.dingtalk.com/microapp/bgb96b
->odoo-DingDing模块组成：  
+> 另外： **因模块功能不完善和存在bug，仅用于个人测试使用，禁止将本模块用于商用！！！若需要商用请联系作者定制开发或授权使用！！！**
+> 
+> 在使用本模块前，请先将钉钉中你创建的E应用或微应用权限放开和配置出口ip，得到钉钉应用的 **AppKey**和 **AppSecret**， 至于钉钉后台中的配置请参照：>https://open-doc.dingtalk.com/microapp/bgb96b 
 >
 > **博客地址： https://sxfblog.com/index.php/archives/371.html**  
-
-> **QQ群：1019231617 若你会odoo开发、或有好的建议欢迎加入群共同努力完善集成钉钉的功能**
-
-> **因模块功能不完善和存在bug，仅用于个人测试使用，请勿将本模块用于商用！！！**
+> **QQ群：1019231617 欢迎加入群进行钉钉集成Odoo开发、共同努力完善集成钉钉的功能**
+> 
+> 安装完成后需要设置用户权限才能看到基础配置项
 
 -**主要功能：**
