+++ ali_dindin/models/__init__.py
@@ -2,8 +2,9 @@
 
 from . import systemc_conf
 from . import res_config_settings
-from . import ali_dindin_util
+from . import dingding_token
 from . import department
 from . import employee
 from . import partner
+from . import synchronous
 from . import feedback
