+++ ali_dindin/models/partner.py
@@ -140,101 +140,6 @@ class ResPartner(models.Model):
         except ReadTimeout:
             raise UserError("上传至钉钉超时！")
 
-    # 在联系人看板、列表中点击同步时执行同步方法
-    @api.model
-    def synchronous_dingding_res_partner(self):
-        """
-        在联系人看板、列表中点击同步时执行同步方法
-        :return:
-        """
-        # 同步联系人标签
-        result = self.synchronous_dingding_category()
-        if not result.get('state'):
-            return result
-        # 同步联系人
-        return self.synchronous_dingding_partner()
-
-    @api.model
-    def synchronous_dingding_category(self):
-        logging.info(">>>同步钉钉联系人标签start")
-        url = self.env['ali.dindin.system.conf'].search([('key', '=', 'listlabelgroups')]).value
-        token = self.env['ali.dindin.system.conf'].search([('key', '=', 'token')]).value
-        headers = {'Content-Type': 'application/json'}
-        data = {
-            'size': 100,
-            'offset': 0
-        }
-        result = requests.post(url="{}{}".format(url, token), headers=headers, data=json.dumps(data), timeout=15)
-        result = json.loads(result.text)
-        if result.get('errcode') == 0:
-            category_list = list()
-            for res in result.get('results'):
-                for labels in res.get('labels'):
-                    category_list.append({
-                        'name': labels.get('name'),
-                        'din_id': labels.get('id'),
-                        'din_color': res.get('color'),
-                        'din_category_type': res.get('name'),
-                    })
-            for category in category_list:
-                res_category = self.env['res.partner.category'].sudo().search([('din_id', '=', category.get('din_id'))])
-                if res_category:
-                    res_category.sudo().write(category)
-                else:
-                    self.env['res.partner.category'].sudo().create(category)
-            return {'state': True}
-        else:
-            logging.info(">>>获取联系人标签失败，原因为:{}".format(result.get('errmsg')))
-            return {'state': False, 'msg': "获取联系人标签失败，原因为:{}".format(result.get('errmsg'))}
-
-    @api.model
-    def synchronous_dingding_partner(self):
-        logging.info(">>>同步钉钉联系人列表start")
-        url = self.env['ali.dindin.system.conf'].search([('key', '=', 'extcontact_list')]).value
-        token = self.env['ali.dindin.system.conf'].search([('key', '=', 'token')]).value
-        headers = {'Content-Type': 'application/json'}
-        data = {
-            'size': 100,
-            'offset': 0
-        }
-        result = requests.post(url="{}{}".format(url, token), headers=headers, data=json.dumps(data), timeout=15)
-        result = json.loads(result.text)
-        if result.get('errcode') == 0:
-            for res in result.get('results'):
-                # 获取标签
-                label_list = list()
-                for label in res.get('label_ids'):
-                    category = self.env['res.partner.category'].sudo().search(
-                        [('din_id', '=', label)])
-                    if category:
-                        label_list.append(category[0].id)
-                data = {
-                    'name': res.get('name'),
-                    'function': res.get('title'),
-                    'category_id': [(6, 0, label_list)],  # 标签
-                    'din_userid': res.get('userid'),  # 钉钉用户id
-                    'comment': res.get('remark'),  # 备注
-                    'street': res.get('address'),  # 地址
-                    'mobile': res.get('mobile'),  # 手机
-                    'phone': res.get('mobile'),  # 电话
-                    'din_company_name': res.get('company_name'),  # 钉钉公司名称
-                }
-                # 获取负责人
-                if res.get('follower_user_id'):
-                    follower_user = self.env['hr.employee'].sudo().search(
-                        [('din_id', '=', res.get('follower_user_id'))])
-                    data.update({'din_employee_id': follower_user[0].id if follower_user else ''})
-                # 根据userid查询联系人是否存在
-                partner = self.env['res.partner'].sudo().search(['|', ('din_userid', '=', res.get('userid')), ('name', '=', res.get('name'))])
-                if partner:
-                    partner.sudo().write(data)
-                else:
-                    self.env['res.partner'].sudo().create(data)
-            return {'state': True}
-        else:
-            logging.info("获取联系人列表失败，原因为:{}".format(result.get('errmsg')))
-            return {'state': False, 'msg': "获取联系人列表失败，原因为:{}".format(result.get('errmsg'))}
-
 
 # 未使用，但是不能删除，因为第一个版本创建的视图还存在
 class DinDinSynchronous(models.TransientModel):
