--- ali_dindin/views/asset.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <template id="ali_dingding_synchronous_template" name="DingDing Synchronous Template" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <script type="text/javascript" src="/ali_dindin/static/src/js/res_partner.js"></script>
-            <script type="text/javascript" src="/ali_dindin/static/src/js/employee.js"></script>
-            <script type="text/javascript" src="/ali_dindin/static/src/js/department.js"></script>
-        </xpath>
-    </template>
-
-</odoo>
