+++ ali_dindin/views/department.xml
@@ -6,14 +6,16 @@
         <field name="model">hr.department</field>
         <field name="inherit_id" ref="hr.view_department_form"/>
         <field name="arch" type="xml">
-             <xpath expr="//sheet" position="before">
-                <header>
-                    <button name="create_ding_department" type="object" string="上传至钉钉" class="oe_highlight"
-                                confirm="确认要将此部门上传至钉钉吗？若不是新创建的部门则无法上传!"/>
-                    <button name="update_ding_department" type="object" string="更新至钉钉" class="oe_highlight"
-                                confirm="系统将新的信息上传至钉钉,确认?"/>
-                </header>
-            </xpath>
+            <button name="toggle_active" position="before">
+                <field name="dingding_type" invisible="1"/>
+                <button name="create_ding_department" string="上传至钉钉" type="object" class="oe_stat_button"
+                        icon="fa-hand-scissors-o" attrs="{'invisible':[('dingding_type','=', 'yes')]}"
+                        groups="ali_dindin.dd_top_group">
+                </button>
+                <button name="update_ding_department" string="更新至钉钉" type="object" class="oe_stat_button"
+                        icon="fa-hand-peace-o" groups="ali_dindin.dd_update_group">
+                </button>
+            </button>
         </field>
     </record>
