+++ ali_dindin/views/synchronous.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="dingding_bash_data_synchronous_form_view" model="ir.ui.view">
+        <field name="name">dingding.bash.data.synchronous.form</field>
+        <field name="model">dingding.bash.data.synchronous</field>
+        <field name="arch" type="xml">
+            <form>
+                <div class="alert alert-info">
+                    <p>选择需要同步的钉钉数据，同步数据量较大的话，等待时间可能会很久额~~</p>
+                </div>
+                <div class="oe_title">
+                    <div name="options">
+                        <div>
+                            <field name="employee"/>
+                            <label for="employee"/>
+                        </div>
+                        <div>
+                            <field name="department"/>
+                            <label for="department"/>
+                        </div>
+                        <div>
+                            <field name="partner"/>
+                            <label for="partner"/>
+                        </div>
+                    </div>
+                </div>
+                <footer>
+                    <button string="开始同步" name="start_synchronous_data" type="object" class="oe_highlight"/>
+                    <button string="不同步、告辞" class="btn btn-default" special="cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record id="dingding_bash_data_synchronous_action" model="ir.actions.act_window">
+        <field name="name">钉钉基础数据同步</field>
+        <field name="res_model">dingding.bash.data.synchronous</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+
+    <menuitem name="数据同步" id="dingding_bash_data_synchronous_menu" parent="ali_dindin.bash_menu" sequence="1"
+              action="dingding_bash_data_synchronous_action" groups="ali_dindin.dd_synchronous_group"/>
+
+    <menuitem name="部门" id="dingding_employee_menu"  parent="ali_dindin.bash_menu" sequence="5"
+          groups="hr.group_hr_user"
+          action="hr.open_module_tree_department"/>
+
+    <menuitem name="员工" id="dingding_department_menu" parent="ali_dindin.bash_menu" sequence="6"
+          groups="hr.group_hr_manager,hr.group_hr_user,base.group_user"
+          action="hr.open_view_employee_list_my"/>
+
+    <menuitem name="联系人" id="dingding_partner_menu" parent="ali_dindin.bash_menu" sequence="7"
+          groups="base.group_user"
+          action="base.action_partner_form"/>
+
+</odoo>
