+++ dindin_callback/views/call_back_list.xml
@@ -39,7 +39,7 @@
               id="dindin_callbacke_main_menu"
               parent="ali_dindin.main_menu"
               sequence="50"
-              groups="ali_dindin.group_ali_dindin_manager"
+              groups="ali_dindin.dd_setting_group"
     />
 
     <menuitem id="dindin_users_callback_listmenu" name="回调类型列表"
